+++ test/router.spec.js
@@ -1,25 +1,23 @@
-import { onNavigate } from '../src/Router';
+import { onNavigate, routes } from '../src/Router';
 
 jest.mock('../src/lib/imports/firebase-imports.js');
 
-console.log(onNavigate);
+const viewTest = () => {
+  const testContainer = document.createElement('section');
+  testContainer.innerHTML = 'I am a view test!';
 
-const routes = [
-  {
-    path: '/ruta-1',
-    template: '<div>ruta 1</div>',
-  },
-  {
-    path: '/ruta-2',
-    template: '<div>ruta 2</div>',
-  },
-];
+  return testContainer;
+};
+
+routes['/ruta-1'] = viewTest;
+
 describe('Router', () => {
   it('onNavigate carga la ruta correcta ', () => {
-    expect(onNavigate(routes.path[0])).toBe('/ruta-1');
+    onNavigate('/ruta-1');
+    expect(window.location.pathname).toBe('/ruta-1');
   });
   it('onNavigate carga la vista correcta ', () => {
-    document.body.innerHTML = routes.template[1];
-    expect(onNavigate(routes.path[0])).toBe(document.body.innerHTML = routes[0].template);
+    onNavigate('/ruta-1');
+    expect(document.getElementById('root').firstChild.innerHTML).toBe('I am a view test!');
   });
 });
